Refactor to package
